Relocate resolver controller to shared utilities
Move createResolverController from src/plugin/src/shared to src/shared/src/utils so resolver plumbing lives with the shared option helpers and future consumers no longer import through plugin internals.
