+++ src/plugin/src/options/core-option-overrides.js
@@ -1,5 +1,5 @@
 import { hasOwn } from "../shared/index.js";
-import { createResolverController } from "../shared/resolver-controller.js";
+import { createResolverController } from "@prettier-plugin-gml/shared/utils/resolver-controller.js";
 import {
     TRAILING_COMMA,
     assertTrailingCommaValue
